Maybe I'm done
I think I'm done, but the algorithm is a little bit unstable in its results. It seems it is too sensitive to the initial values, but I don't know how to fix it . The in-built R command has the option of trying a given number of possible first values and picks the best one, but I have no idea on how it does so. My attempts at it failed miserably, and I haven't been able to find anything on google on how to mimic whatever R's kmeans command does with regards to this.

Either way, besides that pretty major issue, I am pretty much done.